fix(permissions): move assertCanPostInGroup to server-only module
